Added explicit casts near format specifications.
git-svn-id: https://aug.svn.sourceforge.net/svnroot/aug/trunk@68 a3a97f7c-8e16-0410-9893-1d5f7d0f8f6e

+++ src/c/augmar/header_.c
@@ -204,7 +204,7 @@ aug_setvalue_(aug_seq_t seq, struct aug_info_* info, size_t ord,
     if (ord >= info->fields_) {
 
         aug_seterrinfo(__FILE__, __LINE__, AUG_SRCLOCAL, AUG_EEXIST,
-                       AUG_MSG("field '%d' does not exist"), ord);
+                       AUG_MSG("field '%d' does not exist"), (int)ord);
         return -1;
     }
 
@@ -369,7 +369,7 @@ aug_valuebyord_(aug_seq_t seq, const struct aug_info_* info, size_t ord,
     if (ord >= info->fields_) {
 
         aug_seterrinfo(__FILE__, __LINE__, AUG_SRCLOCAL, AUG_EEXIST,
-                       AUG_MSG("field '%d' does not exist"), ord);
+                       AUG_MSG("field '%d' does not exist"), (int)ord);
         return NULL;
     }
