+++ src/c/augsys/debug.c
@@ -18,7 +18,7 @@ dumpclient_(void* userdata, size_t size)
     _ASSERTE(_CrtIsMemoryBlock(userdata, (unsigned int)size, NULL, &file,
                                &line));
     aug_warn("%s(%d) : Memory leak detected at 0x%p, %d byes long",
-             file, line, userdata, size);
+             file, (int)line, userdata, (int)size);
 }
 
 AUGSYS_EXTERN void
