+++ src/c/augsys/posix/mmap.c
@@ -39,7 +39,7 @@ toprot_(int* to, int from)
 
  fail:
     aug_seterrinfo(__FILE__, __LINE__, AUG_SRCLOCAL, AUG_EINVAL,
-                   AUG_MSG("invalid protection flags '%d'"), from);
+                   AUG_MSG("invalid protection flags '%d'"), (int)from);
     return -1;
 }
 
@@ -51,7 +51,7 @@ verify_(size_t size, size_t offset, size_t len)
 
     if (!size || size < (offset + len)) {
         aug_seterrinfo(__FILE__, __LINE__, AUG_SRCLOCAL, AUG_EINVAL,
-                       AUG_MSG("invalid file map size '%d'"), size);
+                       AUG_MSG("invalid file map size '%d'"), (int)size);
         return -1;
     }
 
@@ -59,7 +59,7 @@ verify_(size_t size, size_t offset, size_t len)
 
     if (offset && (offset % aug_granularity())) {
         aug_seterrinfo(__FILE__, __LINE__, AUG_SRCLOCAL, AUG_EINVAL,
-                       AUG_MSG("invalid file map offset '%d'"), offset);
+                       AUG_MSG("invalid file map offset '%d'"), (int)offset);
         return -1;
     }
